Enum s3::Region

pub enum Region {

    UsEast1,
    UsEast2,
    UsWest1,
    UsWest2,
    CaCentral1,
    AfSouth1,
    ApEast1,
    ApSouth1,
    ApNortheast1,
    ApNortheast2,
    ApNortheast3,
    ApSoutheast1,
    ApSoutheast2,
    CnNorth1,
    CnNorthwest1,
    EuNorth1,
    EuCentral1,
    EuCentral2,
    EuWest1,
    EuWest2,
    EuWest3,
    MeSouth1,
    SaEast1,
    DoNyc3,
    DoAms3,
    DoSgp1,
    DoFra1,
    Yandex,
    WaUsEast1,
    WaUsEast2,
    WaUsWest1,
    WaEuCentral1,
    R2 {
        account_id: String,
    },
    Custom {
        region: String,
        endpoint: String,
    },

}

AWS S3 region identifier, passing in custom values is also possible, in that case it is up to you to pass a valid endpoint, otherwise boom will happen :)

Serde support available with the serde feature

Example

use std::str::FromStr;
use awsregion::Region;

// Parse from a string
let region: Region = "us-east-1".parse().unwrap();

// Choose region directly
let region = Region::EuWest2;

// Custom region requires valid region name and endpoint
let region_name = "nl-ams".to_string();
let endpoint = "https://s3.nl-ams.scw.cloud".to_string();
let region = Region::Custom { region: region_name, endpoint };

Variants

UsEast1

us-east-1

UsEast2

us-east-2

UsWest1

us-west-1

UsWest2

us-west-2

CaCentral1

ca-central-1

AfSouth1

af-south-1

ApEast1

ap-east-1

ApSouth1

ap-south-1

ApNortheast1

ap-northeast-1

ApNortheast2

ap-northeast-2

ApNortheast3

ap-northeast-3

ApSoutheast1

ap-southeast-1

ApSoutheast2

ap-southeast-2

CnNorth1

cn-north-1

CnNorthwest1

cn-northwest-1

EuNorth1

eu-north-1

EuCentral1

eu-central-1

EuCentral2

eu-central-2

EuWest1

eu-west-1

EuWest2

eu-west-2

EuWest3

eu-west-3

MeSouth1

me-south-1

SaEast1

sa-east-1

DoNyc3

Digital Ocean nyc3

DoAms3

Digital Ocean ams3

DoSgp1

Digital Ocean sgp1

DoFra1

Digital Ocean fra1

Yandex

Yandex Object Storage

WaUsEast1

Wasabi us-east-1

WaUsEast2

Wasabi us-east-2

WaUsWest1

Wasabi us-west-1

WaEuCentral1

Wasabi eu-central-1

R2

Fields

account_id: String

Custom region

Custom

Fields

region: String

endpoint: String

Implementations

impl Region

pub fn endpoint(&self) -> String

pub fn scheme(&self) -> String

pub fn host(&self) -> String

pub fn from_env( region_env: &str, endpoint_env: Option<&str> ) -> Result<Region, RegionError>

pub fn from_default_env() -> Result<Region, RegionError>

Attempts to create a Region from AWS_REGION and AWS_ENDPOINT environment variables