Crate s3

Simple access to Amazon Web Service’s (AWS) Simple Storage Service (S3)

Re-exports

pub use awscreds as creds;

pub use awsregion as region;

pub use bucket::Bucket;

pub use bucket::Tag;

pub use bucket_ops::BucketConfiguration;

Modules

bucket

bucket_ops

command

deserializer

error

request

request_trait

serde_types

signing

Implementation of AWS V4 Signing

utils

Enums

Region

AWS S3 region identifier, passing in custom values is also possible, in that case it is up to you to pass a valid endpoint, otherwise boom will happen :)