Crate s3

Simple access to Amazon Web Service's (AWS) Simple Storage Service (S3)

Modules

signing
Implementation of AWS V4 Signing

Structs

Bucket
Object holding info about an S3 bucket which provides easy access to S3 operations.

CommonPrefix
CommonPrefix is used to group keys

Credentials
AWS access credentials: access key, secret key, and optional token.

ListBucketResult
The parsed result of a s3 bucket listing

Object
An individual object in a ListBucketResult

Owner
Owner information for the object

S3RegionParseError
Error raised when a string cannot be parsed to a valid Region.

Enums

Region
AWS S3 region identifier

S3Error
Errors raised by the libray when performing generic operations.

Type Definitions

Headers
Collection of HTTP headers sent to S3 service, in key/value format.

Query
Collection of HTTP query parameters sent to S3 service, in key/value format.

S3Result
Generic return type of S3 functions.