Struct s3::bucket::Bucket

pub struct Bucket {
    pub name: String,
    pub region: Region,
    pub credentials: Credentials,
    pub extra_headers: Headers,
    pub extra_query: Query,
}

Example

use s3::bucket::Bucket;
use s3::credentials::Credentials;

let bucket_name = "rust-s3-test";
let region = "us-east-1".parse().unwrap();
let credentials = Credentials::new("access_key", "secret_key", None);

let bucket = Bucket::new(bucket_name, region, credentials);

Fields
name: String region: Region credentials: Credentials extra_headers: Headers extra_query: Query
Methods

impl Bucket

fn new(name: &str, region: Region, credentials: Credentials) -> Bucket

Instantiate a new Bucket.

Example

use s3::bucket::Bucket;
use s3::credentials::Credentials;

let bucket_name = "rust-s3-test";
let region = "us-east-1".parse().unwrap();
let credentials = Credentials::new("access_key", "secret_key", None);

let bucket = Bucket::new(bucket_name, region, credentials);

fn get(&self, path: &str) -> S3Result<(Vec<u8>, u32)>

Gets file from an S3 path.

Example:

use s3::bucket::Bucket;
use s3::credentials::Credentials;

let bucket_name = "rust-s3-test";
let region = "us-east-1".parse().unwrap();
let credentials = Credentials::new("access_key", "secret_key", None);
let bucket = Bucket::new(bucket_name, region, credentials);

let (data, code) = bucket.get("/test.file").unwrap();
println!("Code: {}\nData: {:?}", code, data);

fn delete(&self, path: &str) -> S3Result<(Vec<u8>, u32)>

Delete file from an S3 path.

Example:

use s3::bucket::Bucket;
use s3::credentials::Credentials;

let bucket_name = &"rust-s3-test";
let region = "us-east-1".parse().unwrap();
let credentials = Credentials::new("access_key", "secret_key", None);
let bucket = Bucket::new(bucket_name, region, credentials);

let (_, code) = bucket.delete("/test.file").unwrap();
assert_eq!(204, code);

fn put(
&self,
path: &str,
data: &[u8],
content_type: &str
) -> S3Result<(Vec<u8>, u32)>

Put into an S3 bucket.

Example

use s3::bucket::Bucket;
use s3::credentials::Credentials;

let bucket_name = &"rust-s3-test";
let aws_access = &"access_key";
let aws_secret = &"secret_key";

let bucket_name = &"rust-s3-test";
let region = "us-east-1".parse().unwrap();
let credentials = Credentials::new("access_key", "secret_key", None);
let bucket = Bucket::new(bucket_name, region, credentials);

let content = "I want to go to S3".as_bytes();
let (_, code) = bucket.put("/test.file", content, "text/plain").unwrap();
assert_eq!(201, code);

fn list(
&self,
prefix: &str,
delimiter: Option<&str>
) -> S3Result<(ListBucketResult, u32)>

List the contents of an S3 bucket.

Example

use std::str;
use s3::bucket::Bucket;
use s3::credentials::Credentials;
let bucket_name = &"rust-s3-test";
let aws_access = &"access_key";
let aws_secret = &"secret_key";

let bucket_name = &"rust-s3-test";
let region = "us-east-1".parse().unwrap();
let credentials = Credentials::new("access_key", "secret_key", None);
let bucket = Bucket::new(bucket_name, region, credentials);

let (list, code) = bucket.list("/", Some("/")).unwrap();
assert_eq!(200, code);
println!("{:?}", list);

fn name(&self) -> &str

Get a reference to the name of the S3 bucket.

fn host(&self) -> &str

Get a reference to the hostname of the S3 API endpoint.

fn region(&self) -> Region

Get the region this object will connect to.

fn access_key(&self) -> &str

Get a reference to the AWS access key.

fn secret_key(&self) -> &str

Get a reference to the AWS secret key.

fn token(&self) -> Option<&str>

Get a reference to the AWS token.

fn credentials(&self) -> &Credentials

Get a reference to the full Credentials object used by this Bucket.

fn set_credentials(&mut self, credentials: Credentials) -> Credentials

Change the credentials used by the Bucket, returning the existing credentials.

fn add_header(&mut self, key: &str, value: &str)

Add an extra header to send with requests to S3.

Add an extra header to send with requests. Note that the library already sets a number of headers - headers set with this method will be overridden by the library headers: * Host * Content-Type * Date * Content-Length * Authorization * X-Amz-Content-Sha256 * X-Amz-Date

fn extra_headers(&self) -> &Headers

Get a reference to the extra headers to be passed to the S3 API.

fn extra_headers_mut(&mut self) -> &mut Headers

Get a mutable reference to the extra headers to be passed to the S3 API.

fn add_query(&mut self, key: &str, value: &str)

Add an extra query pair to the URL used for S3 API access.

fn extra_query(&self) -> &Query

Get a reference to the extra query pairs to be passed to the S3 API.

fn extra_query_mut(&mut self) -> &mut Query

Get a mutable reference to the extra query pairs to be passed to the S3 API.