Enum s3::region::Region

pub enum Region {
    UsEast1,
    UsEast2,
    UsWest1,
    UsWest2,
    CaCentral1,
    ApSouth1,
    ApNortheast1,
    ApNortheast2,
    ApNortheast3,
    ApSoutheast1,
    ApSoutheast2,
    CnNorth1,
    CnNorthwest1,
    EuCentral1,
    EuWest1,
    EuWest2,
    EuWest3,
    SaEast1,
    DoNyc3,
    DoAms3,
    DoSgp1,
    Custom(String),
}

AWS S3 region identifier, passing in custom values is also possible, in that case it is up to you to pass a valid endpoint, otherwise boom will happen :)

Example

use std::str::FromStr;
use s3::region::Region;

// Parse from a string
let region: Region = "us-east-1".parse().unwrap();

// Choose region directly
let region = Region::EuWest2;

Variants
UsEast1

us-east-1

UsEast2

us-east-2

UsWest1

us-west-1

UsWest2

us-west-2

CaCentral1

ca-central-1

ApSouth1

ap-south-1

ApNortheast1

ap-northeast-1

ApNortheast2

ap-northeast-2

ApNortheast3

ap-northeast-3

ApSoutheast1

ap-southeast-1

ApSoutheast2

ap-southeast-2

CnNorth1

cn-north-1

CnNorthwest1

cn-northwest-1

EuCentral1

eu-central-1

EuWest1

eu-west-1

EuWest2

eu-west-2

EuWest3

eu-west-3

SaEast1

sa-east-1

DoNyc3

Digital Ocean nyc3

DoAms3

Digital Ocean ams3

DoSgp1

Digital Ocean sgp1

Custom(String)

Custom region

Methods

impl Region

pub fn endpoint(&self) -> String

pub fn scheme(&self) -> String

pub fn host(&self) -> String