Struct s3::serde_types::ListBucketResult

pub struct ListBucketResult {
    pub name: String,
    pub next_marker: Option<String>,
    pub delimiter: Option<String>,
    pub max_keys: i32,
    pub prefix: String,
    pub marker: Option<String>,
    pub encoding_type: Option<String>,
    pub is_truncated: bool,
    pub next_continuation_token: Option<String>,
    pub contents: Vec<Object>,
    pub common_prefixes: Option<Vec<CommonPrefix>>,
}

The parsed result of a s3 bucket listing

Fields
name: String

Name of the bucket.

next_marker: Option<String>

When the response is truncated (that is, the IsTruncated element value in the response is true), you can use the key name in this field as a marker in the subsequent request to get next set of objects. Amazon S3 lists objects in UTF-8 character encoding in lexicographical order.

delimiter: Option<String>

A delimiter is a character you use to group keys.

max_keys: i32

Sets the maximum number of keys returned in the response body.

prefix: String

Limits the response to keys that begin with the specified prefix.

marker: Option<String>

Indicates where in the bucket listing begins. Marker is included in the response if it was sent with the request.

encoding_type: Option<String>

Specifies the encoding method to used

is_truncated: bool

Specifies whether (true) or not (false) all of the results were returned. If the number of results exceeds that specified by MaxKeys, all of the results might not be returned.

next_continuation_token: Option<String> contents: Vec<Object>

Metadata about each object returned.

common_prefixes: Option<Vec<CommonPrefix>>

All of the keys rolled up into a common prefix count as a single return when calculating the number of returns.