Module s3::signing

Implementation of AWS V4 Signing

Constants

FRAGMENT

FRAGMENT_SLASH

Functions

authorization_header

Generate the AWS authorization header.

authorization_query_params_no_sig

canonical_header_string

Generate a canonical header string from the provided headers.

canonical_query_string

Generate a canonical query string from the query pairs in the given URL.

canonical_request

Generate a canonical request.

canonical_uri_string

Generate a canonical URI string from the given URL.

flatten_queries

scope_string

Generate an AWS scope string.

signed_header_string

Generate a signed header string from the provided headers.

signing_key

Generate the AWS signing key, derived from the secret key, date, region, and service name.

string_to_sign

Generate the “string to sign” - the value to which the HMAC signing is applied to sign requests.

uri_encode

Encode a URI following the specific requirements of the AWS service.

Type Definitions

HmacSha256